Refactored the friend fetching and updating logic to a custom hook out of the Friends component
